Add files via upload
public.rb and README.md are not files that I have made. They were provided by the instructors and have only been added as reference to maze.rb. Any attempt at copying or infringing upon any of these 3 files may result in civil action, but especially with the README.md and public.rb. They are to be used solely as references to the coding ability of Yashaswi Sharma